Move some of the node-related functions to the internal/nodes module rather than keeping them in helpers
