+++ datajunction-server/datajunction_server/api/nodes.py
@@ -17,16 +17,11 @@ from sqlalchemy.sql.operators import is_
 from starlette.requests import Request
 
 from datajunction_server.api.helpers import (
-    activate_node,
-    deactivate_node,
     get_catalog_by_name,
     get_column,
     get_node_by_name,
     get_node_namespace,
-    hard_delete_node,
     raise_if_node_exists,
-    revalidate_node,
-    validate_node_data,
 )
 from datajunction_server.api.namespaces import create_node_namespace
 from datajunction_server.api.tags import get_tags_by_name
@@ -51,18 +46,23 @@ from datajunction_server.internal.access.authorization import (
 )
 from datajunction_server.internal.nodes import (
     _create_node_from_inactive,
+    activate_node,
     copy_to_new_node,
     create_cube_node_revision,
     create_node_revision,
+    deactivate_node,
     get_column_level_lineage,
     get_node_column,
+    hard_delete_node,
     remove_dimension_link,
+    revalidate_node,
     save_column_level_lineage,
     save_node,
     set_node_column_attributes,
     update_any_node,
     upsert_complex_dimension_link,
 )
+from datajunction_server.internal.validation import validate_node_data
 from datajunction_server.models import access
 from datajunction_server.models.attribute import AttributeTypeIdentifier
 from datajunction_server.models.dimensionlink import (
